requirements..txt
asyncio
nltk
python-dotenv
telethon